requirements.txt
flet>=0.27.6
qrcode>=8.1
Pillow>=11.2.1
pyzbar>=0.1.9
python-dateutil>=2.8.2 